Migrates pull.monitor.Stats to new plugin layout
