+++ pnp/plugins/pull/monitor/__init__.py
@@ -1,0 +1,7 @@
+"""Monitoring related plugins."""
+
+from .stats import Stats
+
+__all__ = [
+    'Stats'
+]
