+++ pnp/plugins/pull/simple/count.py
@@ -1,4 +1,4 @@
-"""Pull: Count."""
+"""Pull: simple.Count."""
 
 import sys
 import warnings
