+++ pnp/plugins/pull/simple/repeat.py
@@ -1,4 +1,5 @@
-"""Pull: Repeat."""
+"""Pull: simple.Repeat."""
+
 import warnings
 from typing import Any, Optional
