+++ pnp/plugins/pull/simple/run_once.py
@@ -1,4 +1,4 @@
-"""Pull: RunOnce."""
+"""Pull: simple.RunOnce."""
 
 from typing import Any
